added rejecting functionality
ability to reject contirbutions and give a reason. The reasons are taken from taxonomies where you specifiy a taxonomy called Reject and then lists them as options. It records this in the new field reject as a number so if there is a number in reject it is rejected and the number ciorresponds to the reason chosen.